GamuGatewaySMS/requirements.txt
python-gammu
flask
flask-restx
Flask-HTTPAuth
pyopenssl